refactor code to get weather using location navigator
