--- src/style.scss
@@ -1,3 +1,0 @@
-.red {
-  color: red;
-}
